Update the Zandbox, Curve, book, and some tests

+++ zinc-const/src/base.rs
@@ -3,13 +3,13 @@
 //!
 
 /// The binary number base.
-pub const BINARY: usize = 2;
+pub const BINARY: u32 = 2;
 
 /// The octal number base.
-pub const OCTAL: usize = 8;
+pub const OCTAL: u32 = 8;
 
 /// The decimal number base.
-pub const DECIMAL: usize = 10;
+pub const DECIMAL: u32 = 10;
 
 /// The hexadecimal number base.
-pub const HEXADECIMAL: usize = 16;
+pub const HEXADECIMAL: u32 = 16;
