+++ zinc-const/src/zandbox.rs
@@ -21,4 +21,4 @@ pub static CONTRACT_QUERY_URL: &str = "/api/v1/contract/query";
 pub static CONTRACT_CALL_URL: &str = "/api/v1/contract/call";
 
 /// The ETH balance value exponent.
-pub const ETH_BALANCE_EXPONENT: usize = 18;
+pub const ETH_BALANCE_EXPONENT: u32 = 18;
